[MIG]product_code_unique: migration from 11 to 12
