+++ product_code_unique/__manifest__.py
@@ -4,16 +4,12 @@
 {
     "name": "Unique Product Internal Reference",
     "summary": "Set Product Internal Reference as Unique",
-    "version": "11.0.1.0.0",
+    "version": "12.0.1.0.0",
     "license": "AGPL-3",
     "author": "Open Source Integrators, Odoo Community Association (OCA)",
     "category": "Product",
     "website": "https://github.com/OCA/product-attribute",
     "depends": ["product"],
-    "data": [
-    ],
-    'demo': [
-    ],
     "pre_init_hook": 'pre_init_product_code',
     "installable": True,
 }
